app2/render-map-a/requirements.txt
folium
Pillow
mplleaflet
matplotlib

